Move ProposalIter to ProposalStore
Moves the `ProposalIter` store up to the level of the `ProposalStore`;
this is similar to the `CircuitStore`. Pushing the creation of the
iterator up to the proposal store will facilitate easier filtering of
proposals, and it reduces complexity at the lower levels.
